Extracted test_suite_filters
